Realizados testes e atualizaçoes de respansividade e busca

+++ src/main/java/com/dev/gestorNF/business/VendedorService.java
@@ -5,6 +5,7 @@ import com.dev.gestorNF.business.dto.out.VendedorDTOResponse;
 import com.dev.gestorNF.business.mapper.VendedorConverter;
 import com.dev.gestorNF.infrastructure.entity.out.UsuarioEntity;
 import com.dev.gestorNF.infrastructure.entity.out.VendedorEntity;
+import com.dev.gestorNF.infrastructure.repository.NotaFiscalRepository;
 import com.dev.gestorNF.infrastructure.repository.UsuarioRepository;
 import com.dev.gestorNF.infrastructure.repository.VendedorRepository;
 import com.dev.gestorNF.infrastructure.security.JwtUtil;
@@ -21,6 +22,7 @@ public class VendedorService {
     private final JwtUtil jwtUtil;
     private final UsuarioRepository usuarioRepository;
     private final VendedorRepository vendedorRepository;
+    private final NotaFiscalRepository notaFiscalRepository;
 
     public VendedorDTOResponse cadastroVendedor(String token, VendedorDTORequest vendedorDTORequest) {
 
@@ -41,15 +43,31 @@ public class VendedorService {
         return vendedorConverter.paraVendedorDTOResponse(vendedorSalvo);
     }
 
-    public void deletaVendedor(String token ,Long idVendedor) {
+    public void deletaVendedor(String token, Long idVendedor) {
 
         String email = jwtUtil.extrairEmailToken(token.substring(7));
 
         UsuarioEntity usuarioEntity = usuarioRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("Email não encontrado " + email));
+                .orElseThrow(() ->
+                        new RuntimeException("Email não encontrado " + email));
 
-        VendedorEntity vendedorEntity = vendedorRepository.findByIdVendedorAndUsuarioId(idVendedor,usuarioEntity.getId())
-                .orElseThrow(() -> new RuntimeException("Vendedor não encontrado"));
+        VendedorEntity vendedorEntity =
+                vendedorRepository.findByIdVendedorAndUsuarioId(
+                        idVendedor,
+                        usuarioEntity.getId()
+                ).orElseThrow(() ->
+                        new RuntimeException("Vendedor não encontrado"));
+
+        boolean possuiNotas =
+                !notaFiscalRepository
+                        .findByVendedorIdVendedor(idVendedor)
+                        .isEmpty();
+
+        if (possuiNotas) {
+            throw new RuntimeException(
+                    "Não é possível excluir este vendedor porque existem notas fiscais vinculadas a ele."
+            );
+        }
 
         vendedorRepository.delete(vendedorEntity);
     }
